added notes-collection and refactored to use it
