+++ lib/notes.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Request = require('./model/notes-schema.js');
+const modelCollection = require('./model/notes-collection.js');
 
 class Note {
   constructor (options) {
@@ -23,43 +23,18 @@ class Note {
   }
 
   add(){
-    if (this.payload != null && this.category != null){
-      const newRequest = new Request({text: this.payload, category: this.category});
-      return newRequest.save()
-      .then(results => console.log(`Adding note: ${this.payload}`))
-      .catch(err => console.log(err));
-    }
-    console.log('You must add the note text and a category.');
+    modelCollection.create(this.payload, this.category);
   }
     
   list(){ 
-    let findThis = {};
-    if (this.category){
-      findThis = {category: this.category};      
-    }
-    return Request.find(findThis)
-      .then(results => {
-        results.forEach((result) => {
-          console.log(`
-          ${result.text}
-          Category: ${result.category}  ID: ${result._id}
-          ----------------------------------------------------`)
-        })
-      })
-      .catch(err => console.log(err));
-
+    modelCollection.get(this.category);
   }
 
   delete(){
-    return Request.deleteOne({_id: this.payload})
-    .then(results => console.log(`Deleted: ${this.payload}` ))
-    .catch(err => console.log(err));
-
+    modelCollection.delete(this.payload);
   }
 
 }
-
-
   
 
 module.exports = Note;
